Improve testability and feedback on ImportErrors from chatroom demo
